split out routes
